refacto(webapp): create participantlistiem composant
