+++ webapp/src/pages/applications/Organization/Gestion/Jobs/List/ParticipantsTable.tsx
@@ -1,29 +1,15 @@
-import { FC } from 'react';
 import {
-  Tooltip,
-  IconButton,
   Table,
   TableBody,
   TableCell,
   TableHead,
   TableRow,
   TableContainer,
-  Typography,
-  useTheme,
   TableProps,
 } from '@mui/material';
-import AccountBoxIcon from '@mui/icons-material/AccountBox';
-import getFormatLocalDate from 'src/utils/getFormatLocalDate';
-import { useNavigate, useParams } from 'react-router';
+import ParticipantListItem from './subComponents/ParticipantListItem';
 
-const ParticipantsTable: FC<TableProps<any>> = ({
-  items: users, 
-}) => {
-
-  const { id } = useParams();
-  const theme = useTheme();
-  const navigate = useNavigate();
-
+function ParticipantsTable({items}: TableProps<any>) {
   return (
       <TableContainer>
         <Table>
@@ -36,58 +22,12 @@ const ParticipantsTable: FC<TableProps<any>> = ({
             </TableRow>
           </TableHead>
           <TableBody>
-            {users.map((user) => {
+            {items.map((item) => {
               return (
-                <TableRow hover key={user.participant.id}>
-                  <TableCell>
-                    <Typography
-                      variant="body1"
-                      fontWeight="bold"
-                      color="text.primary"
-                      gutterBottom
-                      noWrap
-                    >
-                      {user.participant.lastname}
-                    </Typography>
-                  </TableCell>
-                  <TableCell>
-                  <Typography
-                      variant="body1"
-                      fontWeight="bold"
-                      color="text.primary"
-                      gutterBottom
-                      noWrap
-                    >
-                      {user.participant.firstname}
-                    </Typography>
-                  </TableCell>
-                  <TableCell>
-                  <Typography
-                      variant="body1"
-                      fontWeight="bold"
-                      color="text.primary"
-                      gutterBottom
-                      noWrap
-                    >
-                      {getFormatLocalDate(user.createdAt)}
-                    </Typography>
-                  </TableCell>
-                  <TableCell align="right">
-                    <Tooltip title="Voir le profil" arrow>
-                      <IconButton
-                        onClick={() => navigate(`/association/gestion/missions/${id}/participant/${user.participant.id}`)}
-                        sx={{
-                          '&:hover': { background: theme.colors.error.lighter },
-                          color: theme.palette.error.main
-                        }}
-                        color="inherit"
-                        size="small"
-                      >
-                        <AccountBoxIcon fontSize="small" />
-                      </IconButton>
-                    </Tooltip>
-                  </TableCell>
-                </TableRow>
+                <ParticipantListItem
+                  key={item.id}
+                  user={item}
+                />
               );
             })}
           </TableBody>
